backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
Werkzeug==2.3.7
PyYAML==6.0.1
python-dotenv==1.0.0
redis==5.0.0
gunicorn==21.2.0